requirements.txt
Django==1.11.16
djangorestframework==3.8.2
factory-boy==2.11.1
pytest==3.8.2
